backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
email-validator
psycopg2-binary
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
alembic==1.12.0